Refactor move pipeline into modular components
